refactor: migrated ItemList as projectItemList in useable as a component
